+++ cognite/client/data_classes/capabilities.py
@@ -1236,6 +1236,7 @@ class WorkflowOrchestrationAcl(Capability):
 
     class Scope:
         All = AllScope
+        DataSet = DataSetScope
 
 
 @dataclass