+++ tests/tests_integration/test_api/test_data_workflows.py
@@ -30,9 +30,11 @@ from cognite.client.utils._text import random_string
 
 @pytest.fixture
 def workflow_list(cognite_client: CogniteClient) -> WorkflowList:
+    data_set = cognite_client.data_sets.list(limit=1)[0]
     workflow1 = WorkflowUpsert(
         external_id="integration_test-workflow1",
         description="This is  workflow for testing purposes",
+        data_set_id=data_set.id,
     )
     workflow2 = WorkflowUpsert(
         external_id="integration_test-workflow2",
@@ -266,9 +268,11 @@ def workflow_scheduled_trigger(cognite_client: CogniteClient, add_multiply_workf
 
 class TestWorkflows:
     def test_upsert_delete(self, cognite_client: CogniteClient) -> None:
+        data_set = cognite_client.data_sets.list(limit=1)[0]
         workflow = WorkflowUpsert(
             external_id="integration_test-test_create_delete" + random_string(5),
             description="This is ephemeral workflow for testing purposes",
+            data_set_id=data_set.id,
         )
         cognite_client.workflows.delete(workflow.external_id, ignore_unknown_ids=True)
 
@@ -279,6 +283,7 @@ class TestWorkflows:
             assert created_workflow.external_id == workflow.external_id
             assert created_workflow.description == workflow.description
             assert created_workflow.created_time is not None
+            assert created_workflow.data_set_id == data_set.id
         finally:
             if created_workflow is not None:
                 cognite_client.workflows.delete(created_workflow.external_id)